Conflicts:
	TastyPangyo/WebContent/restaurant/regist_restaurant_form.jsp
	TastyPangyo/WebContent/restaurant/regist_success.jsp
	TastyPangyo/src/tp/controller/DeleteRestaurantImage.java
	TastyPangyo/src/tp/controller/InsertRestaurantServlet.java

+++ TastyPangyo/src/tp/dao/impl/ReviewDaoImpl.java
@@ -159,7 +159,7 @@ public class ReviewDaoImpl implements ReviewDao{
 	}
 
 	@Override
-	public int selectAverageOfKostarByRestaurantId(SqlSession session, int restaurantId) {
+	public double selectAverageOfKostarByRestaurantId(SqlSession session, int restaurantId) {
 		return session.selectOne(MapperPath("selectAverageOfKostarByRestaurantId"), restaurantId);
 	}
